Updates Staged for 3-5

+++ music_stream/views.py
@@ -2,10 +2,12 @@ from django.shortcuts import render_to_response, render, redirect, get_object_or
 from django.template import RequestContext
 from django.http import HttpResponse
 from django.http import HttpResponseRedirect
+from django.template.defaultfilters import slugify
 from django.core.urlresolvers import reverse
 from django.db.models import Q
 
 from music_stream.models import  Song
+from music_stream.models import Result
 from music_stream.forms import SongForm
 
 import os
@@ -25,6 +27,8 @@ def index(request):
 	upload_url = blobstore.create_upload_url('/')
 	view_url = reverse('music_stream.views.index')
 	if request.user.is_authenticated():
+		songList = Song.objects.filter(owner=request.user.username)
+
 		if request.method == 'POST':
 			if 'upload_submit' in request.POST:
 				form = SongForm(request.POST, request.FILES)
@@ -35,11 +39,19 @@ def index(request):
 				# Redirect to the document list after POST
 					#return HttpResponse("SUCCESS")
 					return HttpResponseRedirect(view_url)
-		else:
-			form = SongForm() # A empty, unbound forms
+			if 'purge_all' in request.POST:
+				for s in songList:
+					blobstore.delete(s.songfile.file.blobstore_info.key())
+				songList.delete()
+			if 'purge_selection' in request.POST:
+				for s in songList:
+					if str(s) in request.POST:
+						blobstore.delete(s.songfile.file.blobstore_info.key())
+						s.delete()
+
+		form = SongForm() # A empty, unbound forms
 
 			# Load documents for the list page
-		songList = Song.objects.filter(owner=request.user.username)
 		upload_url, upload_data = prepare_upload(request, view_url)
 			# Render list page with the documents and the form
 		return render_to_response(
@@ -51,23 +63,40 @@ def index(request):
 		return redirect('accounts/login', request)
 
 def search(request):
-	context_dict = {}
-	context_dict['artists'] = None
-	context_dict['songs'] = None
-	context_dict['albums'] = None
-	context_dict['query'] = None
+	
 
 	if request.user.is_authenticated():
 		if request.method == 'POST':
 			if 'search_submit' in request.POST:
+				context_dict = {}
+				context_dict['artists'] = None
+				context_dict['songs'] = None
+				context_dict['albums'] = None
+				context_dict['query'] = None
 				query = request.POST['query'].strip()
 				if query:
-					artists = Song.objects.filter(owner=request.user.username).filter(artist__icontains=query).order_by().values('artist','artist_slug').distinct()
-					# search for albums matching the query or albums from the artist matching the query
-					albums = Song.objects.filter(owner=request.user.username).filter(Q(album__icontains=query) | Q(artist__icontains=query)).order_by().values('album','album_slug').distinct()
+					query2 = slugify(query)
+					artistz = set([s.artist for s in Song.objects.filter(owner=request.user.username)])
+					for a in artistz.copy():
+						if not(query in a) and not(query in str(slugify(a))):
+							artistz.remove(a)
+					albumz = set([s.album for s in Song.objects.filter(owner=request.user.username)])
+					for a in albumz.copy():
+						if not(query in a) and not(query in str(slugify(a))):
+							albumz.remove(a)
+					songs = set([s for s in Song.objects.filter(owner=request.user.username)])
+					for a in songs.copy():
+						if not(query in a.title) and not(query in str(slugify(a.title))):
+							songs.remove(a)
+					albums = []
+					artists = []
+					for a in albumz:
+						b= Result(item=a, item_slug=slugify(a))
+						albums.append(b)
+					for a in artistz:
+						b= Result(item=a, item_slug=slugify(a))
+						artists.append(b)
 
-
-					songs = Song.objects.filter(owner=request.user.username).filter(Q(title__icontains=query) | Q(artist__icontains=query) | Q(album__icontains=query))
 					context_dict['query'] = query
 					context_dict['artists'] = artists
 					context_dict['songs'] = songs
@@ -76,6 +105,8 @@ def search(request):
 
 				else:
 					return HttpResponseRedirect(reverse('music_stream.views.index'))
+			else:
+				return HttpResponseRedirect(reverse('music_stream.views.index'))
 		else:
 			return HttpResponseRedirect(reverse('music_stream.views.index'))
 
@@ -85,18 +116,42 @@ def search(request):
 
 def album(request,album_name_slug):
 	context_dict = {}
-	songs = Song.objects.filter(owner=request.user.username).filter(album_slug__icontains=album_name_slug)
+	songs = Song.objects.filter(owner=request.user.username).filter(album_slug=album_name_slug)
 	context_dict['songs'] = songs
+	context_dict['self'] = album_name_slug
+
+	if request.method == 'POST':
+		if 'purge_all' in request.POST:
+			for s in songs:
+				blobstore.delete(s.songfile.file.blobstore_info.key())
+			songs.delete()
+		if 'purge_selection' in request.POST:
+			for s in songs:
+				if str(s) in request.POST:
+					blobstore.delete(s.songfile.file.blobstore_info.key())
+					s.delete()
+
 
 	return render(request, 'music_stream/album.html',context_dict)
 
 def artist(request,artist_name_slug):
 	context_dict = {}
-	albums = Song.objects.filter(owner=request.user.username).filter(artist_slug__icontains=artist_name_slug).order_by().values('album','album_slug').distinct()
-	songs = Song.objects.filter(owner=request.user.username).filter(artist_slug__icontains=artist_name_slug)
-	context_dict['albums'] = albums
+	albums = Song.objects.filter(owner=request.user.username).filter(artist_slug=artist_name_slug).order_by().values('album','album_slug')
+	songs = Song.objects.filter(owner=request.user.username).filter(artist_slug=artist_name_slug)
+	
 	context_dict['songs'] = songs
-
+	context_dict['self'] = artist_name_slug
+	context_dict['albums'] = albums
+	if request.method == 'POST':
+		if 'purge_all' in request.POST:
+			for s in songs:
+				blobstore.delete(s.songfile.file.blobstore_info.key())
+			songs.delete()
+		if 'purge_selection' in request.POST:
+			for s in songs:
+				if str(s) in request.POST:
+					blobstore.delete(s.songfile.file.blobstore_info.key())
+					s.delete()
 
 	return render(request, 'music_stream/artist.html',context_dict)
 
@@ -104,17 +159,30 @@ def serve(request, pk):
 	upFile = get_object_or_404(Song, pk=pk)
 	return serve_file(request, upFile.songfile, save_as=True)
 
-def purge(request):
-	songList = Song.objects.filter(owner=request.user.username)
-	for song in songList:
-		blobstore.delete(song.songfile.file.blobstore_info.key())
-		song.delete()
-	return redirect('/')
-
-def purgeall(request):
-	songList = Song.objects.all()
-	songList.delete()
-	for b in blobstore.BlobInfo.all():
-		blobstore.delete(b.key())
-	return redirect('/')
 # Create your views here.
+
+def blob(request):
+	if request.user.is_authenticated():
+		songList = Song.objects.all()
+		error= ''
+		if request.method == 'POST':
+			if request.user.username == 'test':
+				if 'purge_all' in request.POST:
+					for s in songList:
+						blobstore.delete(s.songfile.file.blobstore_info.key())
+					songList.delete()
+				if 'purge_selection' in request.POST:
+					for s in songList:
+						if str(s) in request.POST:
+							blobstore.delete(s.songfile.file.blobstore_info.key())
+							s.delete()
+			else:
+				error= 'You\'re not authorized to do that...'
+		
+		return render_to_response(
+								'music_stream/list.html',
+								{'songList': songList, 'error': error},
+								context_instance=RequestContext(request)
+								)
+	else:
+		return redirect('accounts/login', request)
